Split steps into own files
